Implement impl<T> with static-strings
This is the first step to the conversion of typedecl
to static-strings.

Some basic types were defined, and they have ugly names:
- "ss" refers to static-strings
- "sss" refers to split static-strings

Also, the template impl_proxy<T> was defined to use static-strings
on the types that were already converted.

When the conversion is done, the non-static-string elements
will be removed and their names may be used.